refactor: Issue #219完了 - EnhancedClassificationEngineをClassificationEngineにリネーム
## 変更内容

### Phase 1: スキーマ統合
- classification.tsとenhanced-classification.tsを統一
- すべてのZodスキーマとTypeScript型を単一ファイルに集約
- 後方互換性エイリアスを追加

### Phase 2: エンジンファイルリネーム
- enhanced-engine.ts → engine.ts
- EnhancedClassificationEngine → ClassificationEngine
- createEnhancedClassificationEngine → createClassificationEngine
- EnhancedTaskScore → TaskScore
- EnhancedBatchResult → BatchResult

### Phase 3: 全体的なインポート更新
- 10個のファイルのインポートパスを更新
- 全てのクラス参照と関数呼び出しを新しい名前に変更
- テストファイルも完全に更新

## 影響範囲
- ✅ 全てのテストが通過
- ✅ TypeScript型チェックOK
- ✅ ESLintエラーなし
- ✅ Prettier形式適用済み
