+++ src/components/classification/ClassificationDetails.astro
@@ -6,7 +6,7 @@
  * confidence levels, reasoning, and metadata
  */
 import ConfidenceIndicator from './ConfidenceIndicator.astro';
-import type { EnhancedIssueClassification } from '../../lib/schemas/enhanced-classification';
+import type { EnhancedIssueClassification } from '../../lib/schemas/classification';
 
 export interface Props {
   classification: EnhancedIssueClassification;
