+++ src/lib/utils/search.ts
@@ -6,7 +6,7 @@
  */
 
 import type { Issue } from '../schemas/github';
-import { createEnhancedClassificationEngine } from '../classification/enhanced-engine';
+import { createClassificationEngine } from '../classification/engine';
 
 export interface SearchFilters {
   state?: 'open' | 'closed' | 'all';
@@ -346,7 +346,7 @@ export async function searchIssuesWithClassification(
       | undefined;
 
     if (options.sortBy === 'priority') {
-      const engine = await createEnhancedClassificationEngine({
+      const engine = await createClassificationEngine({
         owner: 'nyasuto',
         repo: 'beaver',
       });
